fix(decompiler): preserve compiler-lowered semantics
Split AST lifting paths and harden control flow, capture naming, and table reconstruction against compiler-emitted bytecode. Cache AST rewrite analysis and time each pass to reduce GAMECORE runtime. Add compiler-backed semantic regressions and record the remaining table and local-limit gaps.

+++ Fission.Decompiler/include/Deserializer.hpp
@@ -122,7 +122,7 @@ struct DeserializedFunction {
     std::uint8_t uTypeVersion;
     std::uint8_t uBytecodeVersion;
 
-    std::uint8_t bytecodeId{};
+    std::uint32_t bytecodeId{};
     std::uint8_t maxstacksize{};
     std::uint8_t numparams{};
     std::uint8_t nups{};
